split app-shell into client compoent
this allows the user-layout to pass clerk public key into the provider
